The battery was drained too fast due AUX power rail not deactivated
The code reseted the VCC_ENA_PIN state, pulling it to high when going to
sleep. I noticed the battery was draining too fast, ideally the board
should consume a few micro amperes when going to deep sleep.

After viewing the battery burn graph something didn't made sense, so I
power profiled the board and I found that in deep sleep mode the board
was consuming around 60 - 70 mA.

I traced this to the GPIO that controlled the V_AUX rail being held high
when deep sleep.

The fix was to add gpio_hold_en + gpio_deep_sleep_hold_en before goint
to sleep.

After the firmware fix the board consumes an average of 703 mA and some
spikes of 4mA that need to be investigated.